+++ pyrandall/executors/broker_kafka.py
@@ -1,6 +1,5 @@
 from pyrandall.kafka import KafkaConn
-from pyrandall.types import Assertion
-
+from pyrandall.types import Assertion, ExecutionMode, UnorderedDiffAssertion
 from .common import Executor
 
 
@@ -11,8 +10,12 @@ class BrokerKafka(Executor):
         self.spec = spec
 
     def execute(self, reporter):
-        # assume execution mode is producing for now (simulate test cases)
-        spec = self.spec
+        if self.execution_mode is ExecutionMode.SIMULATING:
+            self.simulate(self.spec, reporter)
+        elif self.execution_mode is ExecutionMode.VALIDATING:
+            self.validate(self.spec, reporter)
+
+    def simulate(self, spec, reporter):
         kafka = KafkaConn()
         kafka.init_producer()
 
@@ -24,6 +27,22 @@ class BrokerKafka(Executor):
                 kafka.produce_message(spec.topic, event)
                 send += 1
             a.actual_value = send
+
+    def validate(self, spec, reporter):
+        kafka = KafkaConn()
+        kafka.check_connection()
+        consumed = kafka.consume(spec.topic, spec.assertions.get("timeout_after", 2.0))
+        with Assertion(
+            "total_events", spec.assertions, "total amount of received events", reporter
+        ) as a:
+            # should not be needed to keep track here
+            # assertions.append(a)
+            a.actual_value = len(consumed)
+
+        with UnorderedDiffAssertion(
+            "unordered", spec.assertions, "unordered events", reporter
+        ) as a:
+            a.actual_value = consumed
 
     def represent(self):
         return (