+++ pyrandall/kafka.py
@@ -18,7 +18,12 @@ class ConsumerState(Enum):
     TIMEOUT_SET = 2
 
 
+class KafkaSetupError(Exception):
+    pass
+
+
 class KafkaConn:
+
     def __init__(self):
         self.consume_lock = ConsumerState.PARTITIONS_UNASSIGNED
 
@@ -26,7 +31,8 @@ class KafkaConn:
     # removes lock for actual consumption
     def callback_on_assignment(self, consumer, partitions):
         self.consume_lock = ConsumerState.PARTITIONS_ASSIGNED
-        log.info(f"Assignment: {partitions}")
+        offsets = consumer.committed(partitions)
+        log.info(f"Assignment: {offsets}")
 
     def check_connection(self):
         def check_callback(error, event):
@@ -36,7 +42,7 @@ class KafkaConn:
                         "This Timout might indicate the broker is down or connection is misconfigured"
                     )
                 log.error(f"Error while producing initial msg: {error}")
-                sys.exit(1)
+                raise KafkaSetupError()
 
         config = ConfigFactory(kafka_client="producer").config
         config["delivery.timeout.ms"] = "3000"  # 3 seconds
@@ -50,7 +56,7 @@ class KafkaConn:
         else:
             log.info(
                 f"Event produced, topic: {event.topic()}, \
-                     partition: {event.partition()}"
+                    partition: {event.partition()}"
             )
 
     def produce_message(self, topic, body, headers=None, partition_key=None):
@@ -92,23 +98,23 @@ class KafkaConn:
         kafka_config_consumer = ConfigFactory(kafka_client="consumer")
         config = kafka_config_consumer.config
         log.info("kafka config for consume %s", config)
-        kcons = Consumer(config)
+        consumer = Consumer(config)
 
         events = []
 
         start_time = time.monotonic()
         timeout_start_time = start_time
-        timeout_consumer = 60.0
+        timeout_consumer = 10.0
 
         # actual consumer starts now
         # subscribe to 1 or more topics and define the callback function
         # callback is only received after consumer.consume() is called!
-        kcons.subscribe([topic], on_assign=self.callback_on_assignment)
-        log.info("Waiting for partition assignment ... (timeout at 60 seconds")
+        consumer.subscribe([topic], on_assign=self.callback_on_assignment)
+        log.info(f"Waiting for partition assignment ... (timeout at {timeout_consumer} seconds")
         try:
             while (time.monotonic() - timeout_start_time) < timeout_consumer:
                 # start consumption
-                messages = kcons.consume(timeout=0.1)
+                messages = consumer.consume(timeout=0.1)
                 # check for partition assignment
                 if self.consume_lock == ConsumerState.PARTITIONS_UNASSIGNED:
                     # this should not happen but we are not 100% sure
@@ -120,6 +126,7 @@ class KafkaConn:
                 # and reset the start time from which to determine timeout
                 # violation
                 elif self.consume_lock == ConsumerState.PARTITIONS_ASSIGNED:
+                    
                     timeout_start_time = time.monotonic()
                     timeout_consumer = topic_timeout
 
@@ -134,8 +141,14 @@ class KafkaConn:
                                 partition: {msg.partition()}, \
                                 topic: {msg.topic()}"
                         )
+                        # TODO: allow assertions to be on message headers etc.
+                        # events.append({
+                        #     "key": msg.key,
+                        #     "headers": msg.headers,
+                        #     "value": msg.value()
+                        # })
                         events.append(msg.value())
-
+            # only executed when while condition becomes false
             else:
                 # at the end check if the partition assignment was achieved
                 if self.consume_lock != ConsumerState.TIMEOUT_SET:
@@ -146,7 +159,7 @@ class KafkaConn:
             pass
 
         finally:
-            kcons.close()
+            consumer.close()
 
         end_time = time.monotonic()
         log.debug(f"this cycle took: {(end_time - start_time)} seconds")