+++ pyrandall/spec.py
@@ -1,4 +1,5 @@
 import os
+import re
 
 import jsonschema
 import yaml
@@ -116,8 +117,6 @@ class ScenarioGroup(object):
         out = []
         item = data["simulate"]
         if item["adapter"] == "requests/http":
-            # TODO: look into removing this state
-            self.simulate_adapter = Adapter.REQUESTS_HTTP
             for spec in item["requests"]:
                 o = self.build_simulate_request_events_spec(spec)
                 out.append(o)
@@ -140,7 +139,7 @@ class ScenarioGroup(object):
             o = self.build_simulate_request(spec, fpath)
             out.append(o)
         del spec["assert_that_responded"]
-        return RequestEventsSpec(requests=out)
+        return RequestEventsSpec(requests=out, adapter=Adapter.REQUEST_HTTP_EVENTS)
 
     def build_simulate_request(self, spec, fpath):
         if self.default_request_url is None:
@@ -160,6 +159,7 @@ class ScenarioGroup(object):
             url=join_urlpath(self.default_request_url, spec.get("path", None)),
             headers=request.get("headers", {}),
             body=request.get("body", None),
+            adapter=None
         )
 
     def build_simulate_broker_spec(self, spec):
@@ -170,6 +170,7 @@ class ScenarioGroup(object):
         assertions = {"events_produced": {"equals_to": len(events)}}
         return BrokerKafkaSpec(
             execution_mode=ExecutionMode.SIMULATING,
+            adapter=Adapter.BROKER_KAFKA,
             events=events,
             assertions=self.flatten_assertions(Adapter.BROKER_KAFKA, assertions),
             topic=spec["topic"],
@@ -181,22 +182,18 @@ class ScenarioGroup(object):
     def build_validate_tasks(self, data):
         out = []
         item = data["validate"]
-        mode = ExecutionMode.VALIDATING
         if item["adapter"] == "requests/http":
-            # look into removing state over lists
-            self.validate_adapter = Adapter.REQUESTS_HTTP
             for spec in item["requests"]:
-                o = self.build_validate_request_spec(mode, spec)
+                o = self.build_validate_request_spec(spec)
                 out.append(o)
 
         if item["adapter"] == "broker/kafka":
-            self.validate_adapter = Adapter.BROKER_KAFKA
             for spec in item["messages"]:
-                o = self.build_validate_broker_spec(mode, spec)
+                o = self.build_validate_broker_spec(spec)
                 out.append(o)
         return out
 
-    def build_validate_request_spec(self, execution_mode, spec):
+    def build_validate_request_spec(self, spec):
         if self.default_request_url is None:
             raise ValueError(
                 f"self.default_request_url is {self.default_request_url}. "
@@ -208,7 +205,8 @@ class ScenarioGroup(object):
 
         # build according to scenario/v2 schema
         return RequestHttpSpec(
-            execution_mode=execution_mode,
+            execution_mode=ExecutionMode.VALIDATING,
+            adapter=Adapter.REQUESTS_HTTP,
             assertions=assertions,
             method=spec.get("method", "GET"),
             url=join_urlpath(self.default_request_url, spec.get("path", None)),
@@ -216,12 +214,11 @@ class ScenarioGroup(object):
             body=None,
         )
 
-    def build_validate_broker_spec(self, execution_mode, spec):
-        assertions = {}
-        # TODO: build assertions for unordered comparison
-
+    def build_validate_broker_spec(self, spec):
+        assertions = self.broker_assertions(spec)
         return BrokerKafkaSpec(
-            execution_mode=ExecutionMode.SIMULATING,
+            execution_mode=ExecutionMode.VALIDATING,
+            adapter=Adapter.BROKER_KAFKA,
             events=[],
             assertions=assertions,
             topic=spec["topic"],
@@ -236,30 +233,50 @@ class ScenarioGroup(object):
                 out[key] = self.format_equals_to_event_file(
                     adapter, value["equals_to_event"]
                 )
+            if "timeout_after" == key:
+                out[key] = self.convert_timeout(value)
         return out
 
-    # Work in progress
-    # def broker_assertions(self, spec):
-    #     assertions = {}
-    #     if "assert_that_empty" in spec:
-    #         assertions.update(
-    #             self.broker_assertions({"total_events": {"equals_to": 0}})
-    #         )
-    #     elif (
-    #         "assert_that_received" in spec
-    #         and "unordered" in spec["assert_that_received"]
-    #     ):
-    #         assertions.update(
-    #             {
-    #                 "messages": {
-    #                     "events": [
-    #                         self.flatten_assertions(Adapter.BROKER_KAFKA, x)
-    #                         for x in spec["unordered"]
-    #                     ],
-    #                     "sorted": False,
-    #                 }
-    #             }
-    #         )
+    def broker_assertions(self, spec):
+        adapter = Adapter.BROKER_KAFKA
+        assertions = {}
+        if "assert_that_empty" in spec:
+            # alias for more verbose syntax like:
+            # is it enough?
+            assertions.update(
+                self.flatten_assertions({"total_events": {"equals_to": 0}})
+            )
+        elif (
+            "assert_that_received" in spec
+            and "unordered" in spec["assert_that_received"]
+        ):
+            sub = spec["assert_that_received"]
+            assertions.update(self.flatten_assertions(adapter, sub))
+            assertions.update(
+                {
+                    "messages": [
+                        self.flatten_assertions(adapter, {"value": item})
+                        for item in sub["unordered"]
+                    ]
+                }
+            )
+        return assertions
+
+    def convert_timeout(self, timeout):
+        # timeout can be '10s' , '10m', '10ms'
+        ms = re.compile(r"[0-9]*ms$")
+        m = re.compile(r"[0-9]*m$")
+        s = re.compile(r"[0-9]*s$")
+        if ms.match(timeout):
+            return int(timeout[0:-2]) / 1000
+        elif m.match(timeout):
+            return float(timeout[0:-1]) * 60
+        elif s.match(timeout):
+            return float(timeout[0:-1])
+        else:
+            return ValueError(
+                f"timeout {timeout} format is not supported, valid exampes: 10s, 10m, 10ms"
+            )
 
     def parse_http_request_template(self, fpath):
         with open(self.build_event_path(fpath), "r") as f: