+++ tests/conftest.py
@@ -1,6 +1,9 @@
 import os
 
 import vcr
+import pytest
+
+from confluent_kafka.admin import AdminClient, ClusterMetadata
 
 defaults = ["method", "scheme", "host", "port", "path", "query"]
 vcr = vcr.VCR(
@@ -11,3 +14,14 @@ vcr = vcr.VCR(
     match_on=(defaults + ["body", "headers"]),
     path_transformer=vcr.VCR.ensure_suffix(".yaml"),
 )
+
+
+@pytest.fixture
+def kafka_cluster_info() -> ClusterMetadata:
+    # this fixture can be used to assert the connection is configured
+    # correctly and functioning
+    kafka = os.environ.get("KAFKA_BOOTSTRAP_SERVERS", "127.0.0.1:9092")
+    admin = AdminClient({"bootstrap.servers": kafka})
+    cluster = admin.list_topics()
+    assert ('{1: BrokerMetadata(1, %s)}' % kafka) == str(cluster.brokers)
+    return cluster